// frontend/src/types/category.ts
export type CategoryIcons =
  | 'briefcase-business'
  | 'car-front'
  | 'heart'
  | 'piggy-bank'
  | 'shopping-cart'
  | 'ticket'
  | 'tool-case'
  | 'utensilis'
  | 'paw-print'
  | 'house'
  | 'gift'
  | 'dumbbell'
  | 'book-open'
  | 'baggage-claim'
  | 'mailbox'
  | 'receip-text';

export type CategoryTones =
  | 'green'
  | 'blue'
  | 'red'
  | 'purple'
  | 'orange'
  | 'pink'
  | 'yellow';